+++ src/pages/generate/components/steps/Step1HouseInfo.css.ts
